+++ pkg/dkwatcher/dkwatcher.go
@@ -10,8 +10,8 @@ import (
 	"golang.org/x/xerrors"
 
 	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infra/dreamkast"
+	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infra/infra"
 	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infra/sharedmem"
-	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infrastructure/db"
 	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/metrics"
 	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/model"
 	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/utils"
@@ -58,7 +58,7 @@ func Run(ctx context.Context, conf Config) error {
 		return err
 	}
 
-	redisClient, err := db.NewRedisClient(conf.RedisHost)
+	redisClient, err := infra.NewRedisClient(conf.RedisHost)
 	if err != nil {
 		return err
 	}
@@ -87,7 +87,7 @@ func Run(ctx context.Context, conf Config) error {
 
 func procedure(ctx context.Context,
 	dkClient dreamkast.Client, mw sharedmem.WriterIface, mr sharedmem.ReaderIface,
-	notificationEventSendChan chan<- model.CurrentAndNextTalk, redisClient *db.RedisClient,
+	notificationEventSendChan chan<- model.CurrentAndNextTalk, redisClient *infra.RedisClient,
 ) error {
 	rootLogger := utils.GetLogger(ctx)
