+++ pkg/infra/infra/redis.go
@@ -1,4 +1,4 @@
-package db
+package infra
 
 import (
 	"context"
