+++ pkg/model/talk.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infrastructure/db"
+	"github.com/cloudnativedaysjp/cnd-operation-server/pkg/infra/infra"
 )
 
 type TalkType int32
@@ -35,8 +35,8 @@ func (ts Talks) IsStartNextTalkSoon(untilNotify time.Duration) bool {
 	return nextTalk.StartAt.Sub(now) <= untilNotify
 }
 
-func (ts Talks) HasNotify(ctx context.Context, rc *db.RedisClient, untilNotify time.Duration) (bool, error) {
-	result := rc.Client.Get(ctx, db.NextTalkNotificationKey)
+func (ts Talks) HasNotify(ctx context.Context, rc *infra.RedisClient, untilNotify time.Duration) (bool, error) {
+	result := rc.Client.Get(ctx, infra.NextTalkNotificationKey)
 	if result.Err() != nil {
 		return false, result.Err()
 	}
